+++ src/main/java/br/com/starti/domain/repository/EmpresaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.starti.domain.entity.Empresa;
 
 @Repository
-public interface EmpresaRepository extends JpaRepository<Empresa,Integer>{
+public interface EmpresaRepository extends JpaRepository<Empresa, Long>{
 
 }
