+++ src/main/java/br/com/starti/domain/repository/VagaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.starti.domain.entity.Vaga; 
  
 @Repository 
-public interface VagaRepository  extends JpaRepository<Vaga, Integer>{ 
+public interface VagaRepository  extends JpaRepository<Vaga, Long>{ 
  
 }
